+++ server/controllers/workout.controller.js
@@ -1,13 +1,14 @@
+const { default: mongoose } = require("mongoose");
 const Workout = require("../models/workout.model");
 
 module.exports.findOneWorkout = (req, res) => {
-    Workout.findOne({ _id: req.params.id })
+    Workout.find({ _id: mongoose.Types.ObjectId(req.params.id),  userId: mongoose.Types.ObjectId(req.params.userId)})
         .then(oneWorkout => res.json({ workout: oneWorkout }))
         .catch(err => res.json({ msg: "An Error Occured", error: err }));
 }
 
 module.exports.findAllWorkouts = (req, res) => {
-    Workout.find()
+    Workout.find({userId: mongoose.Types.ObjectId(req.params.userId)})
         .then(allWorkouts => res.json({ workout: allWorkouts }))
         .catch(err => res.json({ msg: "An Error Occured", error: err }));
 }
@@ -21,7 +22,7 @@ module.exports.createWorkout = (req, res) => {
 }
 
 module.exports.editWorkout = (req, res) => {
-    Workout.findOneAndUpdate({ _id: req.params.id }, req.body, { new: true })
+    Workout.findOneAndUpdate({ _id: mongoose.Types.ObjectId(req.params.id), userId: mongoose.Types.ObjectId(req.params.userId) }, req.body, { new: true })
         .then(editedWorkout => {
             res.json({ workout: editedWorkout })
         })
@@ -29,7 +30,7 @@ module.exports.editWorkout = (req, res) => {
 }
 
 module.exports.deleteWorkout = (req, res) => {
-    Workout.findOneAndDelete({ _id: req.params.id })
+    Workout.findOneAndDelete({ _id: mongoose.Types.ObjectId(req.params.id), userId: mongoose.Types.ObjectId(req.params.userId) })
         .then(deletedWorkout => {
             res.json({ workout: deletedWorkout })
         })
